# feat: improve performance of matrix when multiple selectors are specified