--- a/lib/pact_broker/matrix/query_builder.rb
+++ b/lib/pact_broker/matrix/query_builder.rb
@@ -7,24 +7,8 @@ module PactBroker
 
       def self.provider_matches(query_ids, qualifier)
         {
-          qualify(qualifier, :provider_id) => query_ids.pacticipant_ids,
+          qualify(qualifier, :provider_id) => query_ids.pacticipant_ids
         }
-      end
-
-      def self.provider_or_provider_version_matches_or_pact_unverified(query_ids, provider_version_qualifier = nil, provider_qualifier = nil)
-        ors = provider_or_provider_version_criteria(query_ids, provider_version_qualifier, provider_qualifier)
-
-        # If we have specified any versions, then we need to add an
-        # "OR (provider matches these IDs and provider version is null)"
-        # so that we get a line with blank verification details.
-        if query_ids.pacticipant_version_ids.any?
-          ors << {
-            qualify(provider_qualifier, :provider_id) => query_ids.all_pacticipant_ids,
-            qualify(provider_version_qualifier, :provider_version_id) => nil
-          }
-        end
-
-        Sequel.|(*ors)
       end
 
       def self.provider_or_provider_version_criteria(query_ids, provider_version_qualifier = nil, provider_qualifier = nil)
@@ -34,8 +18,10 @@ module PactBroker
         ors
       end
 
-      def self.consumer_in_pacticipant_ids(query_ids)
-        { consumer_id: query_ids.all_pacticipant_ids }
+      def self.consumer_matches(query_ids, qualifier)
+        {
+          qualify(qualifier, :consumer_id) => query_ids.pacticipant_ids
+        }
       end
 
       def self.consumer_or_consumer_version_matches(query_ids, qualifier)
@@ -59,23 +45,6 @@ module PactBroker
         } , {
           qualify(qualifier, :provider_id) => query_ids.specified_pacticipant_ids
         })
-      end
-
-      # QueryIds is built from a single selector, so there is only one pacticipant_id or pacticipant_version_id
-      def self.consumer_or_consumer_version_or_provider_or_provider_or_provider_version_match(query_ids, pacts_qualifier = :p, verifications_qualifier = :v)
-        ors = if query_ids.pacticipant_version_id
-                [
-                  { Sequel[pacts_qualifier][:consumer_version_id] => query_ids.pacticipant_version_id },
-                  { Sequel[verifications_qualifier][:provider_version_id] => query_ids.pacticipant_version_id }
-                ]
-              else
-                [
-                  { Sequel[pacts_qualifier][:consumer_id] => query_ids.pacticipant_id },
-                  { Sequel[pacts_qualifier][:provider_id] => query_ids.pacticipant_id }
-                ]
-              end
-
-        Sequel.|(*ors)
       end
 
       def self.qualify(qualifier, column)