@bethesque 2023-08-21
Beth in ten years doing the next refactor will thank Beth now for those comments.